LPD-51248 Move general files into main folder
