+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/index.ts
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default as ContentsFDSPropsTransformer} from './FDSPropsTransformer/ContentsFDSPropsTransformer';
-export {default as FilesFDSPropsTransformer} from './FDSPropsTransformer/FilesFDSPropsTransformer';
-export {default as StructuresFDSPropsTransformer} from './FDSPropsTransformer/StructuresFDSPropsTransformer';
-export {default as CategorizationMainView} from './categorization/CategorizationMainView';
-export {default as SpacesSticker} from './components/SpaceSticker';
-export {default as SpacesNavigation} from './spaces_navigation/SpacesNavigation';
+export {default as ContentsFDSPropsTransformer} from './main/FDSPropsTransformer/ContentsFDSPropsTransformer';
+export {default as FilesFDSPropsTransformer} from './main/FDSPropsTransformer/FilesFDSPropsTransformer';
+export {default as StructuresFDSPropsTransformer} from './main/FDSPropsTransformer/StructuresFDSPropsTransformer';
+export {default as CategorizationMainView} from './main/categorization/CategorizationMainView';
+export {default as SpacesSticker} from './main/components/SpaceSticker';
+export {default as SpacesNavigation} from './main/spaces_navigation/SpacesNavigation';
 export {default as StructureBuilder} from './structure_builder/components/StructureBuilder';
