+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/FDSPropsTransformer/cell_renderers/AuthorRenderer.tsx
